+++ FCARDIO.Protocol.USB/OfflinePatrol/OperatedDevice/TriggerDoubleLamp/TriggerDoubleLamp_Parameter.cs
@@ -16,13 +16,13 @@ namespace FCARDIO.Protocol.USB.OfflinePatrol.OperatedDevice.TriggerDoubleLamp
         /// 是否开启
         /// 128 - 绿灯亮 ，8 - 红灯亮 ， 0 - 关闭
         /// </summary>
-        public byte Code;
+        public int Code;
 
         /// <summary>
         /// 初始化参数
         /// </summary>
         /// <param name="code">是否开启 128 - 绿灯亮 ，8 - 红灯亮 ， 0 - 关闭</param>
-        public TriggerDoubleLamp_Parameter(byte code)
+        public TriggerDoubleLamp_Parameter(int code)
         {
             Code = code;
         }
